requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
requests>=2.31.0
python-decouple>=3.8
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mysqlclient==2.2.7
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
